% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_start_of_week.r
\name{calculate_start_of_week}
\alias{calculate_start_of_week}
\title{Caucula o primeiro dia da semana de qualquer data}
\usage{
calculate_start_of_week(date, start = "sun")
}
\arguments{
\item{date}{Data, no formato "yyyy-mm-dd"}

\item{start}{Dia em que a semana começar "sun" ou "mon". Defaults to "sun"}
}
\value{
Primeiro dia da semana a qual a data entrada pertece
}
\description{
Caucula o primeiro dia da semana de qualquer data
}
\author{
Gabriel Maia
}
