% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_maia.R
\name{map_maia}
\alias{map_maia}
\title{Plota o mapa com países pequenos representados por pontos}
\usage{
map_maia(df, region, fill, factor = T, ncolors = 7,
  is_small = "is_small")
}
\arguments{
\item{df}{Data frame a ser plotado}

\item{region}{Coluna compactível com o "region" do map_data}

\item{fill}{Coluna com os valores a serem plotados}
}
\value{
none, plots map
}
\description{
Plota o mapa com países pequenos representados por pontos
}
\author{
Gabriel Maia
}
